create add/update/delete DriverRating api's

+++ database/migrations/2025_12_12_150843_create_driver_ratings_table.php
@@ -26,7 +26,7 @@ return new class extends Migration
                 ->constrained('users')
                 ->cascadeOnDelete();
 
-            $table->unsignedTinyInteger('rating');
+            $table->decimal('rating', 2, 1);
             $table->text('comment')->nullable();
 
             $table->timestamps();
